requirements.txt
numpy
pandas==1.1.3
python-dotenv
mysql-connector-python
pymysql
scikit-learn

-e .